Refactor OSMReader and WaySegmentParser
WaySegmentParser makes most of the OSM reading and got quite complex.

To-Dos:
- Logging
- Correct State Transitions during Parsing
