+++ core/src/main/java/com/graphhopper/reader/osm/OSMParse2.java
@@ -1,0 +1,16 @@
+package com.graphhopper.reader.osm;
+
+import com.graphhopper.reader.ReaderWay;
+
+public class OSMParse2 implements OSMParseInterface {
+	private WayPreprocessor wayPreprocessor;
+	
+	public OSMParse2(WayPreprocessor wayPreprocessor) {
+		this.wayPreprocessor = wayPreprocessor;
+	}
+		
+    @Override
+    public void handleWay(ReaderWay way) {
+    	wayPreprocessor.preprocessWay(way);
+    }
+}
