+++ core/src/test/java/com/graphhopper/reader/osm/OSMReaderTest.java
@@ -587,7 +587,7 @@ public class OSMReaderTest {
         rel.add(new ReaderRelation.Member(ReaderElement.Type.WAY, 2L, "via"));
         rel.add(new ReaderRelation.Member(ReaderElement.Type.WAY, 3L, "to"));
 
-        List<OSMTurnRestriction> osmRel = OSMReader.createTurnRestrictions(rel);
+        List<OSMTurnRestriction> osmRel = RelationHandlerBase.createTurnRestrictions(rel);
         assertEquals(osmRel.get(0).getViaType(), OSMTurnRestriction.ViaType.WAY);
         assertEquals(osmRel.get(0).getViaOSMIds().get(0), 2);
     }
@@ -608,7 +608,7 @@ public class OSMReaderTest {
         viaOSMIds.add(3L);
         viaOSMIds.add(4L);
 
-        List<OSMTurnRestriction> osmRel = OSMReader.createTurnRestrictions(rel);
+        List<OSMTurnRestriction> osmRel = RelationHandlerBase.createTurnRestrictions(rel);
         assertEquals(osmRel.get(0).getViaType(), OSMTurnRestriction.ViaType.MULTI_WAY);
         assertEquals(osmRel.get(0).getViaOSMIds(), viaOSMIds);
     }
@@ -623,7 +623,7 @@ public class OSMReaderTest {
         rel.add(new ReaderRelation.Member(ReaderElement.Type.NODE, 3L, "via"));
         rel.add(new ReaderRelation.Member(ReaderElement.Type.WAY, 4L, "to"));
 
-        List<OSMTurnRestriction> osmRel = OSMReader.createTurnRestrictions(rel);
+        List<OSMTurnRestriction> osmRel = RelationHandlerBase.createTurnRestrictions(rel);
         assertEquals(2, osmRel.size());
 
         assertEquals(1, osmRel.get(0).getOsmIdFrom());
@@ -1021,8 +1021,8 @@ public class OSMReaderTest {
 
     @Test
     public void testFixWayName() {
-        assertEquals("B8, B12", OSMReader.fixWayName("B8;B12"));
-        assertEquals("B8, B12", OSMReader.fixWayName("B8; B12"));
+        assertEquals("B8, B12", WayHandler.fixWayName("B8;B12"));
+        assertEquals("B8, B12", WayHandler.fixWayName("B8; B12"));
     }
 
     private AreaIndex<CustomArea> createCountryIndex() {
